Modified the security-policy configuration to use the new grantAll directive.  This allowa us to eliminate a dependency on other packages with custom permissions. basically, we grant all permissions to managers.
Also moved the include of the config file up to the top-level
directory. This needs tobe included after all of the package
includes. Also, the security policy is different from other packages,
since there must be exactly one security policy plugged in.

Perhaps the meta configuration should be combined with the regular
configuration.


git-svn-id: svn://svn.zope.org/repos/main@24771 62d5b8a3-27da-0310-9561-8e5933582275