Merge from 2.7 branch:
Collector 789:  Zope's transaction behavior flawed.
                                                                                
Historically, errors that made it up to the publisher have run outside the
context of a transaction.  This has caused problems for applications which
expect to be able to write to the database (if only temporarily, before
transaction.abort() is eventually called).
                                                                                
With this patch, we allow the error to execute in the same transaction as
the "main" request, and abort only after the error has executed.



git-svn-id: svn://svn.zope.org/repos/main@24892 62d5b8a3-27da-0310-9561-8e5933582275