Onlinehelp fix. Context help for interfaces and view work.
While traversing to the ++help++ namespace the onlinehelp instance
gets the context of the ++help++ namespace.
This is used in the view to extract the correct help topic.

Changed the way the help topics box is generated. It now
uses a recursivly called page template for rendering the 
links.

added a simple macro to the rotterdam skin, that is
used by the online help.

Added a functional test for the contex help



git-svn-id: svn://svn.zope.org/repos/main@24924 62d5b8a3-27da-0310-9561-8e5933582275