We had an object hub service. The object hub service was responsible for:
- Managing short ids for objects, useful for indexing

- Keeping track of object locations.  This was important when the
  object hub was created, because it wasn't practical to use direct
  object references. No it is, so hub ids are no-longer useful for
  implementing location-independent object references.

The object hub service is dead.

In the future, there will be a utility for use by indexes, that
maintains short ids for objects. Perhaps this will be an indexing id
service.

The index and catalog machinery depend heavily on a facility for
assigning shot ids to objects, currently the hub.  Rather than trying
to remove this dependency now, and ad it back later when we have an
index-id utility, it makes more sense to just remove the index and
catalog facilities for now.

Later, when we create an index-id utility, we'll port the code, which
will still be in the repository, back.



git-svn-id: svn://svn.zope.org/repos/main@24990 62d5b8a3-27da-0310-9561-8e5933582275