Refactoring session machinery
I'm relying on adapters more now that we can do so. RAMSessionDataContainer
appears to be working fine, although as I wanted to behave identically
to the PersistentSessionDataContainer it is fully transactional so
it will give conflict errors if you abuse it. Hopefully I'm not (ab)using
MappingStorage in an unsupported way ;-)

Using sessions from TALES should all fall into place once the
adapter syntax has been nutted out and implemented.



git-svn-id: svn://svn.zope.org/repos/main@25075 62d5b8a3-27da-0310-9561-8e5933582275