Add a cyclic import detector.
Run "utilities/importtool --cyclic-imports somescript [scriptargs]" to run
somescript with the cycle detector enabled.  As soon as a cyclic import is
detected (which will often be well after the import is complete!) an exception
will be raised which reports the problem and the "import stack" that caused
import cycle to be detected.


git-svn-id: svn://svn.zope.org/repos/main@25118 62d5b8a3-27da-0310-9561-8e5933582275