Changes to access logging: - it's called "access" logging, not "hit" logging - keep the "hitlog" configuration section, but generate a   DeprecationWarning telling that the "accesslog" section is preferred
Response to checkin message from revision 25133, which introduced <hitlog>:

I wish I'd had time to review the original introduction of <hitlog>
when it went in; it would not be where it is now.  The changes to
ZConfig will not be eternalized to the ZConfig project; I'll have to
change things around before the second beta.

Hopefully I'll be able to review zope.server.logger at that point as
well.


git-svn-id: svn://svn.zope.org/repos/main@25242 62d5b8a3-27da-0310-9561-8e5933582275