Shameless hack to increase security checking performance
It is much faster to call operators, especially from C, than it is to
call methods.  Now if a checker implements __setitem__, it will be
called rather than check or check_getattr.  Similarly, if a checker 
implements __getitem__, it will be called rather than proxy. Yes, this
is an egregious hack, but it does yield a significant speedup and is
thus worth it. Hopefully, it is well marked.

This change reduces the time required to display a sample contents
page by about 5%.  This percentage will increase as other
optimizations are made and things get faster.


git-svn-id: svn://svn.zope.org/repos/main@25949 62d5b8a3-27da-0310-9561-8e5933582275