Merged from trunk 25905: Removed spurious "exception polling in loop(): " output.
Aded code to catch infinate loops (which occasionally occur for me, 
probably related to hyperthreading) and close all conections in test
servers. This will tend to cause test failures, which is good.

There is probably a bug here, but no one has time to chase it down. :(

We'll just have to limp along until we have time to switch to twisted.


git-svn-id: svn://svn.zope.org/repos/main@26054 62d5b8a3-27da-0310-9561-8e5933582275