Merged from trunk 25920: Changed basic checkers to use dictionaries.
Now when you create checkers, you must pass one or 
two dictionary objects. We used to allow functions 
to be passed that would be called to get the permission
needed to access a name.  It turns out that this generality
wasn't needed or used.  If we need this in the furture, we
can add custom checkers.  For now, we only allow
dictionaries, as that will enable more efficient checker
implementation. 


git-svn-id: svn://svn.zope.org/repos/main@26058 62d5b8a3-27da-0310-9561-8e5933582275