Zope3-dev Collector #139: Memory leak involving buckets and connections
Connection objects were typically immortal because the threaded
transaction manager kept them in an ever-growing lists.  Reworked the
transaction manager internals to use a simple implementation of weak sets
instead.  This plugs all leaks in the test program attached to the
collector report (which was leaking about 100KB/sec on my box).


git-svn-id: svn://svn.zope.org/repos/main@26199 62d5b8a3-27da-0310-9561-8e5933582275