Fix the DAV breakage introduced by r26364.
The PROPFIND view before didn't care whether schema adapters provided
the values as attributes or methods; if they were methods, it called
them.  This is not possible anymore as of r26364, which broke DAV
severely.  This fix turns all those methods into getattr'able
attributes (using properties) and finally provides a test for
DAVSchemaAdapter (this bug did not show up in tests at all).

git-svn-id: svn://svn.zope.org/repos/main@26634 62d5b8a3-27da-0310-9561-8e5933582275