Backport of:
------------------------------------------------------------------------
r26444 | srichter | 2004-07-13 11:01:30 -0400 (Tue, 13 Jul 2004) | 5 
lines

While I was working on functional tests, I found the scenario when this
path is chosen, which is when you get the source (for editing for
example) and macro expansion is turned on. It is perfectly fine to pass
an empty dict at this point, since the template itself is not evaluated.

------------------------------------------------------------------------
------------------------------------------------------------------------
r26445 | srichter | 2004-07-13 11:01:55 -0400 (Tue, 13 Jul 2004) | 2 
lines

Converted XXX to TODO.

------------------------------------------------------------------------



git-svn-id: svn://svn.zope.org/repos/main@27053 62d5b8a3-27da-0310-9561-8e5933582275