Collector #265:  Failure in test_dochttp.py
At least test0001.response contains binary data, so must
be opened in binary mode on Windows.

I'm finding it impossible to add info to the collector issue
(error pages result), so here's what I tried to add there:

"""
Confirmed on WinXP w/ current Zope3 trunk.

Don't know what the test is trying to do, but it's almost certain
that the cause is opening binary files in text mode.  That is, at
least test0001.response is opened in text mode, but contains at least
one largish blob of binary data following GIF89a.

dochttp.py and test_dochttp.py don't have the SVN eol-style:native
property set, so hard to read them here <wink>.  The .response and
.request files don't have eol-style set either, so they're also
treated (by SVN) as binary files.

If I open them in binary mode, the test passes, so I'll check that
in (can't hurt on Linux).

Leaving this open so Jim figures out which of these files he intended
to be text and which binary.
"""


git-svn-id: svn://svn.zope.org/repos/main@27286 62d5b8a3-27da-0310-9561-8e5933582275