Merged from ZopeX3-3.0 branch:
  r27289 | jim | 2004-08-27 10:28:18 -0400 (Fri, 27 Aug 2004) | 14 lines

Added code to create a local grant for the test user.
Later, we need to figure out a way to avoid this depenency on
the security policy.

Added a flag to the http function to disable error handling.  This is
useful for debugging failures.  (Maybe we should make this the
default when a 200 status is expected.)

Added code to unquote the url path. (Perhaps later we
should add logic to utf-8 decode it too.)

Fixed handling of certain headers that don't get an HTTP_ prefix in
the request. This makes xml-rpc requests work


git-svn-id: svn://svn.zope.org/repos/main@27310 62d5b8a3-27da-0310-9561-8e5933582275