Fixed support for CopiedEvent. It wasn't being sent correctly before, as manage_afterClone is only triggered *after* the object is added and thus manage_afterAdd is triggered. manage_afterAdd was assuming manage_afterClone would be triggered *before* the afterAdd. Basically manage_afterClone can not be used to inform manage_afterAdd of the event of a copy.
Instead of relying on manage_afterClone, a hack was added to
manage_afterAdd. An attempt is made to check whether the original
location of the object still has something there. If there is, we assume
the object has just been copied. If not, we'll assume the object has
been moved. This seems to do the trick, even though weird circumstances
can be imaged where moving an object still leaves some object in the original
location for some reason..
-- 
M    tests/test_event.py
M    eventconfigure.py


git-svn-id: svn://svn.zope.org/repos/main@48037 62d5b8a3-27da-0310-9561-8e5933582275