Call close_when_done() rather than close() in FTP data connections.
ftp/server.py: a lot of rearranging was necessary to fix code that 
called close() with arguments.  It's important to send a report at the 
end of FTP data connections, but the old way relied on close() with 
arguments and we sometimes need to call close_when_done() instead of 
close().

dualmodechannel.py: added an assertion that verifies close() is always 
called in asynchronous mode.  See the comment.




git-svn-id: svn://svn.zope.org/repos/main@27460 62d5b8a3-27da-0310-9561-8e5933582275