Check in stuff mistakenly committed from the Zope trunk.
Rev 26842 changed cache-size parameters to type byte-size.
That's a good change, but it was checked in from a wrong place
(from the Zope trunk's copy of ZODB).  Since we want Zope 2.8 to
be usable with ZODB 3.3 too, I need to slam this change into
3.3 now.

Alas, I don't understand how ZODB+ZEO got stitched into the
Zope trunk (I do understand how it got stitched into the
Zope3 trunk, and into the X3 branch), and it looks like that
was left in a poor state.


git-svn-id: svn://svn.zope.org/repos/main@27621 62d5b8a3-27da-0310-9561-8e5933582275