Transaction.begin().
When raising DeprecationWarning, point the warning machinery at
the caller, not at Transaction.begin() itself.  As is (voice
of experience <wink>), these warnings are darned near useless to
track down code that needs to be changed on Zope trunk.  With the
change, the warnings point directly at the deprecated uses.


git-svn-id: svn://svn.zope.org/repos/main@27695 62d5b8a3-27da-0310-9561-8e5933582275