Rewrite more instances of the deprecated Transaction_object.begin().
Caution:  Since no deprecation warnings about Transaction.begin()
were raised by the test suite before this patch, none of this code
is exercised by the test suite -- or it is, but "something"
suppresses the deprecation warnings.  So it's hard to claim that
this patch has been tested; the strongest claim I can make is that
all the unit tests still pass.


git-svn-id: svn://svn.zope.org/repos/main@27703 62d5b8a3-27da-0310-9561-8e5933582275