Brrr.
Weak sets have have pragmatic gotchas, explained in the comments
before the new WeakSet.as_weakref_list() method.  In essence, we
just took all the weak sets of connection objects and changed
everything so that a list of live objects is never materialized
anymore.  Also added new map()-like methods so that clients don't
usually need to be aware of the weakrefs under the covers.


git-svn-id: svn://svn.zope.org/repos/main@28341 62d5b8a3-27da-0310-9561-8e5933582275