Merge rev 28435 from 3.3 branch.
Change ConflictError constructor to stop importing app objects.

When the ConflictError constructor is passed a pickle, extract
the module and class names without loading the pickle.  A ZEO
server doesn't necessarily have the implementation code for
application classes, and when that's so the attempt to raise
ConflictError was itself dying with an ImportError.


git-svn-id: svn://svn.zope.org/repos/main@28436 62d5b8a3-27da-0310-9561-8e5933582275