contents.html no longer breaks when no IAdding is available.
This was pretty involved:

- IAdding did not fully document the methods expected by contents.html, 
and BasicAdding did not implement the required methods.  Therefore, I
combined Adding and BasicAdding into a single class called Adding and 
added the missing methods to the interface.

- nameAllowed and namesAccepted apparently served the same purpose and 
did exactly the same thing, so I changed all references to namesAccepted 
to nameAllowed (nameAllowed won because a unit test for it exists.)

- Added ZCML that takes effect only when running functional tests.

- Added a functional test that verifies contents.html can display the 
contents of a read-only container without breaking.



git-svn-id: svn://svn.zope.org/repos/main@28448 62d5b8a3-27da-0310-9561-8e5933582275