Two error handling corrections:
1. If we can't get the path of an object in order to annotate the 
transaction, we have to swallow the TypeError and not record any path.  
However, a better solution would be for the path discovery code to 
raise something more meaningful than TypeError.

2. When rendering an exception view, call mapply() directly to 
avoid passing positional arguments.  Otherwise, the system breaks 
when trying to report errors that occur in XML-RPC methods that
accept arguments. 


git-svn-id: svn://svn.zope.org/repos/main@28465 62d5b8a3-27da-0310-9561-8e5933582275