Changed CONTENT_VIEW from @@manage to @@SelectedManagementView.html.
@@manage and @@SelectedManagementView.html do the same thing but have 
different permissions applied to them.  @@manage requires the 
ManageContent permission, while the other is public.  Therefore, with 
CONTENT_VIEW set to @@manage, only people with the ManageContent
permission can use the XML tree.  In reality, lots of people without 
ManageContent can use the ZMI, so it's better for links to use 
@@SelectedManagementView.html.


git-svn-id: svn://svn.zope.org/repos/main@28467 62d5b8a3-27da-0310-9561-8e5933582275