Collector 322:  test_bad_import vs Python 2.4c1
Reverted revisions 28420 & 28426.  Those introduced a distinction
in ConfigurationContext.resolve() that depended on Python's long-
time bad behavior of leaving insane module objects behind in
sys.modules after a failing import.  Python 2.4 no longer does
that, so the test depending on that distinction broke.

The intent of the new code was fine, but it can't be achieved in
so simple a way.


git-svn-id: svn://svn.zope.org/repos/main@28480 62d5b8a3-27da-0310-9561-8e5933582275