Fixed a metal bug revealed by the Rotterdam skin.
You should be able to fill a slot defined in a nested define-macro.  
It's unclear how the old code behaved in such a situation, but I fixed 
the bug by removing a strange case in the code.  The code was trying to
prevent slot fillers from being inherited under some special 
case, but the case is apparently undocumented and untested, since all 
the tests still pass once I removed the strange code.  I added a test 
for the bug revealed by Rotterdam.



git-svn-id: svn://svn.zope.org/repos/main@28494 62d5b8a3-27da-0310-9561-8e5933582275