Worked around the multiple user source bug.
When there is more than one principal source, @@grant.html displays a 
search field for each source.  The submitted form then has two values 
for the search field and the publisher interprets the multiple values as 
a list.  The principal search machinery breaks when it gets passed a 
list as a search string.  Workaround: concatenate the list elements to 
form a string.  Not a solution because we still get two fields on the 
page that look exactly alike--very confusing.

The proper solution is probably to change SourceInputWidget to find a 
single widget that will suffice for all sources.  That's Jim's decision, 
though.


git-svn-id: svn://svn.zope.org/repos/main@28495 62d5b8a3-27da-0310-9561-8e5933582275