Got all tests to pass.
I deactivated a couple of module's tests for now, since they depend on 
the correct functioning of the new software-space UI and BBB, which is 
not done yet:

zope.app.component.bbb -- Clearly these tests should not pass until BBB 
is completed.

zope.app.component.browser -- I have not worked on the UI at all yet.

zope.app.module -- I changed the way persistent modules should work, so I 
need to update the code accordingly first to make the tests pass again.

zope.app.presentation -- The registration of page folders heavily 
depended on the old way of the registration framework; I need to redesign 
this to make it work again.

zope.app.pluggableauth -- I could have updated the code to work as a 
utility, but it is deprecated anyways, so I do not want to spend time on 
it. Once I have completed the BBB code, the tests should work just fine.



git-svn-id: svn://svn.zope.org/repos/main@28773 62d5b8a3-27da-0310-9561-8e5933582275