Changed the strategy for handling invalidation of classes. No-longer use setklassstate.  Instead, just call _p_invalidate, as with any other object.  This changes didn't break any tests, so I assume that this was untested. :(
Changed the strategy for invalidating objects.  Non-ghostifiable
objects will load their state when they are invalidated.  We have to
worry about other invalidations that come in while this is happening.
See the comment in _flush_invalidations.

We need to force all invalidations to take this into account, by going
through _flush_invalidations.  I haven't done this yet, but I have
left some XXX comments in places where it needs to be done to remind
myself that this needs to be done.


git-svn-id: svn://svn.zope.org/repos/main@29067 62d5b8a3-27da-0310-9561-8e5933582275