Reversed:
revision 29031:
  Added support for unauthenticated principals.

This was added in preparation for using PAS to get the unauthenticated 
principal to be in the unauthenticated group.  I ended up handling
that group in a way that akes this change innecessary.  While being
able to define an unauthenticated principal in PAS is, theoretically a
useful thing, we have no practical need for it.


git-svn-id: svn://svn.zope.org/repos/main@29074 62d5b8a3-27da-0310-9561-8e5933582275