Changed the PrincipalSource getQueriables method, now we return each utility only once. Changed the render method in QuerySchemaSearchView for rendering zope3.css like html output. Changed the source widget for rendering zope3 CSS conform html output. Changed the getQueriables in authentication.py, ensure that we access the right searchers if we call the getQueriables method from a PrincipalSource vocabulary. Changed test setup for zope.app.authentication.README.txt to placefull, for supporting getUtility context.
TODO:
Perhaps we shold change the source widget that the search forms QuerySchemaSearchView just show up if we 
havn't selected a source item. If we selected a source item we can show just a link for selecting another item.
This whould simplyfie the search result, which in the authentication grant.html form is a bigger table.

git-svn-id: svn://svn.zope.org/repos/main@29096 62d5b8a3-27da-0310-9561-8e5933582275