Merge rev 29194 from ZODB 3.3 branch.
This can be ignored next time a new ZODB is stitched into
Zope3.  This fixes a gross boost in process highwater mark
when running the tests.  After this, running the ZODB/ZEO
--all tests hits a highwater mark of 35MB on WinXP, in one
test, and hovers in the low-to-mid 20s before and after.
Other platforms may or may not behave better or worse.
The highwater mark on these tests was well over 200MB before
this patch.

setUp():  Use a much smaller (than default) ZEO client cache.

The current default is 200MB, and at least testSerialization reads
the entire cache file into one giant string.  That grossly boosts
the process highwater mark when running the tests.  Windows reclaims
the memory after, but other platforms may not.

Cool:  testSerialization was one of the slowest tests before, but
is 100x faster now.



git-svn-id: svn://svn.zope.org/repos/main@29196 62d5b8a3-27da-0310-9561-8e5933582275