zope.app.module tests were shadowed accidentally: the file hierarchy was flattened and tests have been moved up from a subdirectory, but the reference to README.txt was not updated, so the test case would run zope/app/README.txt instead of zope/app/module/README.txt.  I wonder if we do not have a similar problem with other packages.
The tests were quite outdated.  I managed to fix most of the problems, with
two exceptions: module.__name__ is not set, and the import hook does not
work properly for some reason.  I have commented out the corresponding
sections in the test.

The test had implied that module2.Blah() (line #93) should return None, which
does not make sense to me, so I assumed that this was an oversight.  Tests
expect resolve() to return None when the module can not be found, so I updated
resolve() to behave that way instead of propagating an AttributeError.


git-svn-id: svn://svn.zope.org/repos/main@29241 62d5b8a3-27da-0310-9561-8e5933582275