Import hotshot only if profiling is enabled.
It was recently discovered that the standard Python profile module comes
with a licence that is not entirely free.  Some Linux distributions
consequently removed it from the standard Python distribution and moved it into
a separate non-free package.  As a result, 'import hotshot.stats' fails
because it tries to reuse the pstats module from the standard profiler.



git-svn-id: svn://svn.zope.org/repos/main@29259 62d5b8a3-27da-0310-9561-8e5933582275